chore: move hook to it's component directory & extract width resolver
