+++ app/client/src/pages/Editor/WidgetsEditor/components/MainContainerWrapper/MainContainerWrapper.tsx
@@ -22,7 +22,6 @@ import {
   getSelectedAppTheme,
 } from "selectors/appThemingSelectors";
 import { getCanvasWidgetsStructure } from "@appsmith/selectors/entitiesSelector";
-import { useDynamicAppLayout } from "utils/hooks/useDynamicAppLayout";
 import Canvas from "pages/Editor/Canvas";
 import type { AppState } from "@appsmith/reducers";
 import { getIsAnonymousDataPopupVisible } from "selectors/onboardingSelectors";
@@ -30,6 +29,7 @@ import { MainContainerResizer } from "layoutSystems/common/mainContainerResizer/
 import { useMainContainerResizer } from "layoutSystems/common/mainContainerResizer/useMainContainerResizer";
 import { getIsAnvilLayout } from "layoutSystems/anvil/integrations/selectors";
 import { getIsCanvasInitialized } from "selectors/mainCanvasSelectors";
+import { useCanvasWidthAutoResize } from "./hooks";
 
 interface MainCanvasWrapperProps {
   isPreviewMode: boolean;
@@ -108,7 +108,7 @@ const Wrapper = styled.section<{
  * @prop currentPageId, current page id in string
  * @returns
  */
-function MainContainerWrapper(props: MainCanvasWrapperProps) {
+export function MainContainerWrapper(props: MainCanvasWrapperProps) {
   const { isAppSettingsPaneWithNavigationTabOpen, navigationHeight } = props;
   const dispatch = useDispatch();
   const {
@@ -136,7 +136,7 @@ function MainContainerWrapper(props: MainCanvasWrapperProps) {
   const isAnvilLayout = useSelector(getIsAnvilLayout);
 
   const wrapperRef = useRef<HTMLDivElement>(null);
-  useDynamicAppLayout(wrapperRef);
+  useCanvasWidthAutoResize(wrapperRef);
 
   useEffect(() => {
     return () => {
@@ -237,5 +237,3 @@ function MainContainerWrapper(props: MainCanvasWrapperProps) {
     </>
   );
 }
-
-export default MainContainerWrapper;
