+++ app/client/src/pages/Editor/WidgetsEditor/components/MainContainerWrapper/hooks/constants.ts
@@ -1,0 +1,1 @@
+export const RESIZE_DEBOUNCE_THRESHOLD = 100;
