+++ app/client/src/pages/Editor/WidgetsEditor/components/MainContainerWrapper/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useCanvasWidthAutoResize } from "./useCanvasWidthAutoResize";
