+++ app/client/src/pages/Editor/WidgetsEditor/components/MainContainerWrapper/index.ts
@@ -1,0 +1,1 @@
+export { MainContainerWrapper as default } from "./MainContainerWrapper";
