+++ app/client/src/pages/Editor/WidgetsEditor/components/MainContainerWrapper/utils/index.ts
@@ -1,0 +1,1 @@
+export { resolveCanvasWidth } from "./resolveCanvasWidth";
